+++ src/transformer/index.d.ts
@@ -4,8 +4,19 @@
  * @module
  */
 
-import * as comment_parser from 'comment-parser';
 import ts from 'typescript';
+import * as comment_parser from 'comment-parser';
+
+/**
+ * A convenience function to post a debug log message via the `logger` for a Typescript AST node. You may select other
+ * valid log levels. This is handy when debugging AST transformer development.
+ *
+ * @param {ts.Node}  node - Typescript AST node to log.
+ *
+ * @param {'fatal' | 'error' | 'warn' | 'info' | 'debug' | 'verbose' | 'trace'} [logLevel='debug'] Optional alternate
+ *        logging level.
+ */
+declare function logTSNode(node: ts.Node, logLevel?: 'fatal' | 'error' | 'warn' | 'info' | 'debug' | 'verbose' | 'trace'): void;
 
 /**
  * Removes all nodes with the matching JSDoc tags provided. This is useful for handling the `@internal` tag removing
@@ -18,16 +29,26 @@ import ts from 'typescript';
  */
 declare function jsdocRemoveNodeByTags(tags: string | Iterable<string>): ts.TransformerFactory<ts.Bundle | ts.SourceFile>;
 /**
- * Provides a convenient "meta-transformer" that invokes a handler function for each Node w/ the parsed leading
- * comment data for the Node. Only leading block comments are parsed. The `parsed` array is in the data format provided
- * by the `comment-parser` package. For convenience there are `lastComment` and `lastParsed` fields that return the
- * last comment block respectively before the node. Typically, the last comment is the active JSDoc block for a Node.
+ * Provides a convenient "meta-transformer" that invokes a handler function for each Node reducing the boilerplate
+ * required w/ the parsed leading comment data for the Node. Only leading block comments are parsed. The `parsed` array
+ * is in the data format provided by the `comment-parser` package. For convenience there are `lastComment` and
+ * `lastParsed` fields that return the last comment block respectively before the node. Typically, the last comment is
+ * the active JSDoc block for a Node.
  *
- * Note: In the handler return null to remove the Node.
+ * Note: In the `handler` return null to remove the Node. The `postHandler` allows final modification of the SourceFile
+ * after all nodes are visited; return a new SourceFile to update it.
  *
- * @param {(data: { node: ts.Node, sourceFile: ts.SourceFile, comments: string[],
- *         parsed: import('comment-parser').Block[], lastComment: string,
- *          lastParsed: import('comment-parser').Block }) => *}  handler - A function to process JSDoc comments.
+ * @param {((data: {
+ *    node: ts.Node,
+ *    sourceFile: ts.SourceFile,
+ *    comments: string[],
+ *    parsed: import('comment-parser').Block[],
+ *    lastComment: string,
+ *    lastParsed: import('comment-parser').Block
+ * }) => *)}  handler - A function to process AST nodes with JSDoc comments.
+ *
+ * @param {(sourceFile: ts.SourceFile) => ts.SourceFile | undefined} [postHandler] - A function to postprocess the
+ *        source file after all nodes visited. Return an updated SourceFile node.
  *
  * @returns {ts.TransformerFactory<ts.Bundle|ts.SourceFile>} JSDoc custom "meta-transformer".
  */
@@ -38,7 +59,7 @@ declare function jsdocTransformer(handler: (data: {
     parsed: comment_parser.Block[];
     lastComment: string;
     lastParsed: comment_parser.Block;
-}) => any): ts.TransformerFactory<ts.Bundle | ts.SourceFile>;
+}) => any, postHandler?: (sourceFile: ts.SourceFile) => ts.SourceFile | undefined): ts.TransformerFactory<ts.Bundle | ts.SourceFile>;
 /**
  * Returns the leading comment strings for a Node.
  *
@@ -70,4 +91,26 @@ declare function parseLeadingComments(node: ts.Node, sourceFile: ts.SourceFile):
     lastParsed: comment_parser.Block;
 };
 
-export { getLeadingComments, jsdocRemoveNodeByTags, jsdocTransformer, parseLeadingComments };
+/**
+ * Provides a convenient "meta-transformer" that invokes a handler function for each Node reducing the boilerplate
+ * required.
+ *
+ * Note: In the `handler` return null to remove the Node. The `postHandler` allows final modification of the SourceFile
+ * after all nodes are visited; return a new SourceFile to update it.
+ *
+ * @param {((data: {
+ *    node: ts.Node,
+ *    sourceFile: ts.SourceFile,
+ * }) => *)}  handler - A function to process AST nodes.
+ *
+ * @param {(sourceFile: ts.SourceFile) => ts.SourceFile | undefined} [postHandler] - A function to postprocess the
+ *        source file after all nodes visited. Return an updated SourceFile node.
+ *
+ * @returns {ts.TransformerFactory<ts.Bundle|ts.SourceFile>} JSDoc custom "meta-transformer".
+ */
+declare function transformer(handler: (data: {
+    node: ts.Node;
+    sourceFile: ts.SourceFile;
+}) => any, postHandler?: (sourceFile: ts.SourceFile) => ts.SourceFile | undefined): ts.TransformerFactory<ts.Bundle | ts.SourceFile>;
+
+export { getLeadingComments, jsdocRemoveNodeByTags, jsdocTransformer, logTSNode, parseLeadingComments, transformer };