+++ test/postprocess/test.js
@@ -2,6 +2,10 @@ import {
    processInheritDoc,
    PostProcess }           from '../../src/postprocess/index.js';
 
+import { logger }          from '#util';
+
+logger.setLogLevel('verbose');
+
 PostProcess.process({
    filepath: './test/fixture/index.d.ts',
    output: './test/output/index-mod.d.ts',